Move Loadable into wrappers/loadable.js
Loadable is a wrapper we use to load state from a pre-existing deposit flow that a user is continuing. It seems appropriate that it belongs in wrappers/.
